refactor(ChecklistConverter/feat-goal): Checklist에서 엔티티 생성 로직 Converter로 분리 #1
ChecklistConverter 메서드명 의미 전달 명확하도록 수정
